+++ Changelog.md
@@ -8,6 +8,12 @@ The changelog is valid starting with Castberg Project Portal v0.1.0-alpha.
 - [Portal] Change nginx version to fix dynatrace
 - [PipingAndHeatTrace] - 3D for tags
 
+## 1.5.0
+- [PipingAndHeatTrace] New custom group header for HT cables
+- [ReleaseControl] Initial release to test
+- [Workspace] Search in workspace
+- [Scope change] Bugfix
+- [Portal] Help pages
 
 ## 1.4.9 
 - [Workspace] Fix column picker resetting when filtering